fix: avoid triggering multiple animationend events
To prevent the triggering of multiple `animationend` events, a fix has been implemented for combined
animations such as: `fade-in-grow`, `fade-out-shrink` and `slide-up-fade-in`.

Previously, an `animationend` event would be triggered for each individual animation, leading to
unexpected behavior. This fix introduces an `infinite` iteration count in one of the animations,
ensuring that only a single `animationend` event is triggered.